+++ pycatdetector/channels/__init__.py
@@ -3,4 +3,8 @@
 from .AbstractChannel import AbstractChannel
 from .HaGoogleSpeak import HaGoogleSpeak
 from .DiscordWebhook import DiscordWebhook
-__all__ = ["AbstractChannel", "HaGoogleSpeak", "DiscordWebhook"]
+from .BlinkstickSquare import BlinkstickSquare
+__all__ = ["AbstractChannel", 
+           "HaGoogleSpeak", 
+           "DiscordWebhook", 
+           "BlinkstickSquare"]